Failed language server attempt refactor
To summarize a long story I've spent a solid day trying to refactor large parts of the script to support lua language server, only to realize near the end that there is no clean way (or any way really) for it to handle SWEP, ENT and other such tables.

The changes here are the result of me stripping the annotations back out and leaving as much as possible intact. This boils down to every global module being uppercase now, library or not and module() being fully phased out in favor of just normal tables.

There'll probably be further changes to formatting now that I've pushed this through and there might be some regressions as a result of bits that have been skipped over by accident but overall it should provide some benefits to readability in the long run.